New activity_trace_function and show_connections option.
The new activity_trace_function allows to simulate and trace functions
in order to highlight the execution path of activity diagrams with
do_plantuml_activity.

PlantumlArchitecture option show_connections allows to turn on/off
visibility of all connections in architecture diagrams (default = True)

The plantuml architecture connections used to align the layout are back
to invisible.